examples/temporal/basic_workflow/pyproject.toml
[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

[project]
name = "temporal-basic-workflow-example"
version = "0.1.0"
description = "Basic Temporal workflow example for Glean connectors"
requires-python = ">=3.10,<4.0"
dependencies = [
    "glean-indexing-sdk>=0.2.0",
    "temporalio>=1.7.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/basic_workflow"]
